refactor: refactor the UserMenu component
