+++ src/pages/Users/table.tsx
@@ -1,17 +1,10 @@
 import { Avatar } from '@chakra-ui/react';
 import { router } from '@shared/router';
 import { authVar } from '@shared/store/globalAuthState';
-import { MenuRoot } from '@shared/ui/menu';
 import { createColumnHelper } from '@tanstack/react-table';
-import { EllipsisVertical } from 'lucide-react';
+import UsersMenu from '@widgets/ui/users/UsersMenu';
 
-import {
-  StyledAvatar,
-  StyledChvronRight,
-  StyledProfileMenuContent,
-  StyledProfileMenuItem,
-  StyledProfileMenuTrigger,
-} from './users.styles';
+import { StyledAvatar, StyledChvronRight } from './users.styles';
 
 interface Profile {
   id: string;
@@ -78,33 +71,7 @@ export const columns = [
     header: '',
     cell: ({ row }) =>
       authEmail === row.original.email ? (
-        <MenuRoot>
-          <StyledProfileMenuTrigger asChild>
-            <EllipsisVertical />
-          </StyledProfileMenuTrigger>
-          <StyledProfileMenuContent>
-            <StyledProfileMenuItem
-              value="profile"
-              onClick={() => {
-                router.navigate({
-                  to: '/users/$userId',
-                  params: { userId: row.original.id },
-                });
-              }}
-            >
-              Profile
-            </StyledProfileMenuItem>
-            <StyledProfileMenuItem value="update">
-              Update user
-            </StyledProfileMenuItem>
-            <StyledProfileMenuItem
-              value="delete"
-              disabled={row.original.role === 'Employee'}
-            >
-              Delete user
-            </StyledProfileMenuItem>
-          </StyledProfileMenuContent>
-        </MenuRoot>
+        <UsersMenu row={row} />
       ) : (
         <StyledChvronRight
           onClick={() => {
